test_requirements.txt
mock
django-nose
coverage
vcrpy
